+++ packages/plugin-node/src/node.ts
@@ -1,11 +1,16 @@
 import { createPlugin } from "@flint.fyi/core";
 
-import assertStrict from "./rules/assertStrict.js";
-import assertStyles from "./rules/assertStyles.js";
-import blobReadingMethods from "./rules/blobReadingMethods.js";
-import bufferAllocators from "./rules/bufferAllocators.js";
-import eventClasses from "./rules/eventClasses.js";
-import exportsAssignments from "./rules/exportsAssignments.js";
+import assertStrict from "./rules/assertStrict.ts";
+import assertStyles from "./rules/assertStyles.ts";
+import blobReadingMethods from "./rules/blobReadingMethods.ts";
+import bufferAllocators from "./rules/bufferAllocators.ts";
+import consoleSpaces from "./rules/consoleSpaces.ts";
+import eventClasses from "./rules/eventClasses.ts";
+import exportsAssignments from "./rules/exportsAssignments.ts";
+import filePathsFromImportMeta from "./rules/filePathsFromImportMeta.ts";
+import fileReadJSONBuffers from "./rules/fileReadJSONBuffers.ts";
+import nodeProtocols from "./rules/nodeProtocols.ts";
+import processExits from "./rules/processExits.ts";
 
 export const node = createPlugin({
 	name: "Node.js",
@@ -14,7 +19,12 @@ export const node = createPlugin({
 		assertStyles,
 		blobReadingMethods,
 		bufferAllocators,
+		consoleSpaces,
 		eventClasses,
 		exportsAssignments,
+		filePathsFromImportMeta,
+		fileReadJSONBuffers,
+		nodeProtocols,
+		processExits,
 	],
 });