+++ packages/ts/src/rules/emptyDestructures.ts
@@ -1,6 +1,9 @@
-import * as ts from "typescript";
+import type * as ts from "typescript";
 
-import { TypeScriptFileServices, typescriptLanguage } from "../language.js";
+import {
+	type TypeScriptFileServices,
+	typescriptLanguage,
+} from "../language.ts";
 
 export default typescriptLanguage.createRule({
 	about: {