+++ packages/ts/src/rules/forDirections.test.ts
@@ -1,5 +1,5 @@
-import rule from "./forDirections.js";
-import { ruleTester } from "./ruleTester.js";
+import rule from "./forDirections.ts";
+import { ruleTester } from "./ruleTester.ts";
 
 ruleTester.describe(rule, {
 	invalid: [