+++ packages/yaml/src/prepareYamlFile.ts
@@ -1,10 +1,12 @@
-import { LanguageFileDefinition } from "@flint.fyi/core";
-import * as yamlParser from "yaml-unist-parser";
+import type { LanguageFileDefinition } from "@flint.fyi/core";
+import type * as yamlParser from "yaml-unist-parser";
 
-import { parseDirectivesFromYamlFile } from "./directives/parseDirectivesFromYamlFile.js";
+import { parseDirectivesFromYamlFile } from "./directives/parseDirectivesFromYamlFile.ts";
+import type { YamlFileServices } from "./language.ts";
+import type { YamlNodesByName } from "./nodes.ts";
 
 export function prepareYamlFile(
-	languageFile: LanguageFileDefinition,
+	languageFile: LanguageFileDefinition<YamlNodesByName, YamlFileServices>,
 	root: yamlParser.Root,
 	sourceText: string,
 ) {