+++ vitest.config.ts
@@ -2,6 +2,9 @@ import { readdirSync } from "node:fs";
 import { defineConfig } from "vitest/config";
 
 export default defineConfig({
+	resolve: {
+		conditions: ["@flint.fyi/source"],
+	},
 	test: {
 		projects: readdirSync("./packages").map((name) => ({
 			test: {
@@ -11,7 +14,7 @@ export default defineConfig({
 				root: `./packages/${name}`,
 				setupFiles: [
 					"console-fail-test/setup",
-					"../core/lib/ts-patch/install-patch-hooks.js",
+					"@flint.fyi/ts-patch/install-patch-hooks",
 				],
 				testTimeout: 10_000,
 				typecheck: {