+++ packages/plugin-spelling/CHANGELOG.md
@@ -1,5 +1,36 @@
 # @flint.fyi/plugin-spelling
 
+## 0.1.4
+
+### Patch Changes
+
+- 6f6a1d8: fix: added cspell plugin missing config error message and docs update
+- 602c75c: chore: rework packaging with tsdown
+- Updated dependencies [9a8ecc1]
+- Updated dependencies [8b8f8d3]
+- Updated dependencies [ff52cb1]
+- Updated dependencies [602c75c]
+  - @flint.fyi/utils@0.14.0
+  - @flint.fyi/text@0.14.2
+  - @flint.fyi/core@0.18.1
+
+## 0.1.3
+
+### Patch Changes
+
+- d99170f: fix: add missing ("phantom") dependencies to package.jsons
+- 3117eaf: feat: add optional async teardown() for rules
+- Updated dependencies [483ee56]
+- Updated dependencies [d99170f]
+- Updated dependencies [3617e4f]
+- Updated dependencies [f37f0d0]
+- Updated dependencies [5e23e96]
+- Updated dependencies [3117eaf]
+  - @flint.fyi/core@0.17.0
+  - @flint.fyi/rule-tester@0.15.0
+  - @flint.fyi/utils@0.13.3
+  - @flint.fyi/text@0.14.1
+
 ## 0.1.2
 
 ### Patch Changes